Add io config interface
Now tools/config_cores can be used to change the configuration
of io accesses.  Ports 0x0000-0xffff may be configured
to pass through/trap individually, or all ports can
be configured on an individual basis.

This patch also removes some IO vmcalls that are not
necessary for the above configuration options.