Finish EPT trapping configuration
EPT is disabled on VM start-up. It can be turned
on with vmconfig.  Turning it on enables EPT on every core.
The guest physical address space is identity mapped with
2MB pages, each of which are split to 4k-page identity maps
as trap requests come in (as seen in the Bareflank EPT hook
example).

Currently, there are hard-coded limits for the guest address
space, number of simultaneous 4k pages to trap on, and
number of 2MB pages that may be split to 4k identity maps.
These values may need to be tweaked depending on the available
memory of the system.