Linux multi-core EFI
A dirty rendition of booting SMP Linux from EFI. x2APIC is used and
seems to be working fine. Booting into SMP Linux from xAPIC gets the
cores out of boot, but eventually freezes when an interrupt-heavy task is
performed by the user (watching a cat video etc).

Much of the code here is in support for xAPIC, and thus isn't strictly
necessary for x2apic. Note that this patch is based off of
commit 1b43f3e07125b20cb1d of the base hypervisor at
https://github.com/connojd/hypervisor/tree/linux-smp-efi-museum

Signed-off-by: Connor Davis <davisc@ainfosec.com>